refactor(file): restructure filetype package to file package
- Moved classes from `org.js.vento.plugin.filetype` to `org.js.vento.plugin.file` for better structure alignment.
- Renamed classes and adjusted references accordingly (e.g., `VentoFileType` to `FileType`, `VentoFileViewProvider` to `FileViewProvider`).
- Updated `plugin.xml` to reflect new package names for registered extensions.
- Simplified initialization and imports across affected files.
- Adjusted test files to align with new package structure.
