+++ src/main/kotlin/org/js/vento/plugin/actions/ToggleCommentAction.kt
@@ -12,7 +12,7 @@ import com.intellij.openapi.actionSystem.CommonDataKeys
 import com.intellij.openapi.command.WriteCommandAction
 import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.project.Project
-import org.js.vento.plugin.filetype.VentoFileType
+import org.js.vento.plugin.file.VentoFileType
 
 class ToggleCommentAction : AnAction() {
     override fun getActionUpdateThread(): ActionUpdateThread = ActionUpdateThread.BGT
