+++ src/main/kotlin/org/js/vento/plugin/file/TemplateUtil.kt
@@ -1,5 +1,5 @@
 // Copyright 2000-2024 JetBrains s.r.o. and contributors. Use of this source code is governed by the Apache 2.0 license.
-package org.js.vento.plugin.filetype
+package org.js.vento.plugin.file
 
 import com.intellij.injected.editor.VirtualFileWindow
 import com.intellij.lang.Language
