+++ src/main/kotlin/org/js/vento/plugin/highlighting/ScriptContentProvider.kt
@@ -9,7 +9,7 @@ import com.intellij.lang.HtmlScriptContentProvider
 import com.intellij.lexer.Lexer
 import com.intellij.psi.tree.IElementType
 
-class VentoScriptContentProvider : HtmlScriptContentProvider {
+class ScriptContentProvider : HtmlScriptContentProvider {
     override fun getScriptElementType(): IElementType? = null
 
     override fun getHighlightingLexer(): Lexer? = null
