+++ src/test/resources/testdata/echo/ echo string.HTML.txt
@@ -1,0 +1,5 @@
+HtmlFile:a.vto(0,24)
+  PsiElement(HTML_DOCUMENT)(0,0)
+    PsiElement(XML_PROLOG)(0,0)
+      <empty list>
+  PsiElement(VENTO_OUTER_TOKEN)('{{ echo "Hello World" }}')(0,24)
