+++ src/test/resources/testdata/echo/ echo with pipe.HTML.txt
@@ -1,0 +1,23 @@
+HtmlFile:a.vto(0,69)
+  PsiElement(HTML_DOCUMENT)(0,58)
+    PsiElement(VENTO_OUTER_TOKEN)('{{ echo |> md }}')(0,16)
+    PsiWhiteSpace('\n\n')(16,18)
+    PsiElement(XML_PROLOG)(18,18)
+      <empty list>
+    XmlToken:XML_DATA_CHARACTERS('##')(18,20)
+    PsiWhiteSpace(' ')(20,21)
+    XmlToken:XML_DATA_CHARACTERS('Header')(21,27)
+    PsiWhiteSpace('\n\n')(27,29)
+    XmlToken:XML_DATA_CHARACTERS('-')(29,30)
+    PsiWhiteSpace(' ')(30,31)
+    XmlToken:XML_DATA_CHARACTERS('First')(31,36)
+    PsiWhiteSpace(' ')(36,37)
+    XmlToken:XML_DATA_CHARACTERS('item.')(37,42)
+    PsiWhiteSpace('\n')(42,43)
+    XmlToken:XML_DATA_CHARACTERS('-')(43,44)
+    PsiWhiteSpace(' ')(44,45)
+    XmlToken:XML_DATA_CHARACTERS('Second')(45,51)
+    PsiWhiteSpace(' ')(51,52)
+    XmlToken:XML_DATA_CHARACTERS('item.')(52,57)
+    PsiWhiteSpace('\n')(57,58)
+  PsiElement(VENTO_OUTER_TOKEN)('{{ /echo }}')(58,69)
